+++ packages/ub-pub/CHANGELOG.md
@@ -8,13 +8,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
-  - ClientRepository with `__mip_recordhistory_all` misc will bypass a client cache 
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+## [5.9.18] - 2021-11-05
+### Changed
+  - ClientRepository with `__mip_recordhistory_all` misc will bypass a client cache 
 
 ## [5.9.17] - 2021-09-24
 ## [5.9.16] - 2021-09-08